Rename TypedListNode<T> to ListNode<T> and colocate in ListNode.cs
